fix(web): redirects for sign in flow
