+++ apps/nextjs/src/app/auth/sign-in/page.tsx
@@ -1,150 +1,12 @@
-"use client";
+import { redirect } from "next/navigation";
 
-import { useState } from "react";
-import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { useForm } from "@tanstack/react-form";
-import { toast } from "sonner";
-import { z } from "zod";
+import { isAuthenticated } from "~/lib/auth-server";
+import SignInForm from "./sign-in-form";
 
-import { Button } from "@package/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@package/ui/card";
-import { Field, FieldError, FieldLabel } from "@package/ui/field";
-import { Input } from "@package/ui/input";
+export default async function LoginPage() {
+  if (await isAuthenticated()) {
+    redirect("/app");
+  }
 
-import { authClient } from "~/lib/auth-client";
-
-const loginSchema = z.object({
-  email: z.string().email("Please enter a valid email address"),
-  password: z.string().min(1, "Password is required"),
-});
-
-export default function LoginPage() {
-  const router = useRouter();
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const form = useForm({
-    defaultValues: {
-      email: "",
-      password: "",
-    },
-    validators: {
-      onSubmit: loginSchema,
-    },
-    onSubmit: async ({ value: { email, password } }) => {
-      await authClient.signIn.email({
-        email: email,
-        password: password,
-        fetchOptions: {
-          onRequest: () => {
-            setIsSubmitting(true);
-          },
-          onComplete: () => {
-            setIsSubmitting(false);
-          },
-          onSuccess: () => {
-            router.push("/app");
-          },
-          onError: (error) => {
-            toast.error(error.error.message);
-          },
-        },
-      });
-    },
-  });
-
-  return (
-    <div className="flex min-h-screen items-center justify-center p-4">
-      <Card className="w-full max-w-md">
-        <CardHeader>
-          <CardTitle>Welcome back</CardTitle>
-          <CardDescription>
-            Sign in to your Leopard account to continue
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          <form
-            onSubmit={(e) => {
-              e.preventDefault();
-              void form.handleSubmit(e);
-            }}
-            className="space-y-4"
-          >
-            <form.Field
-              name="email"
-              children={(field) => {
-                const isInvalid =
-                  field.state.meta.isTouched && !field.state.meta.isValid;
-                return (
-                  <Field data-invalid={isInvalid}>
-                    <FieldLabel htmlFor={field.name}>Email</FieldLabel>
-                    <Input
-                      id={field.name}
-                      name={field.name}
-                      value={field.state.value}
-                      onBlur={field.handleBlur}
-                      onChange={(e) => field.handleChange(e.target.value)}
-                      aria-invalid={isInvalid}
-                      placeholder="arian@example.com"
-                    />
-                    {isInvalid && (
-                      <FieldError errors={field.state.meta.errors} />
-                    )}
-                  </Field>
-                );
-              }}
-            />
-            <form.Field
-              name="password"
-              children={(field) => {
-                const isInvalid =
-                  field.state.meta.isTouched && !field.state.meta.isValid;
-                return (
-                  <Field data-invalid={isInvalid}>
-                    <FieldLabel htmlFor={field.name}>Password</FieldLabel>
-                    <Input
-                      id={field.name}
-                      type="password"
-                      name={field.name}
-                      value={field.state.value}
-                      onBlur={field.handleBlur}
-                      onChange={(e) => field.handleChange(e.target.value)}
-                      aria-invalid={isInvalid}
-                      placeholder="Enter your password"
-                    />
-                    {isInvalid && (
-                      <FieldError errors={field.state.meta.errors} />
-                    )}
-                  </Field>
-                );
-              }}
-            />
-
-            <Button type="submit" className="w-full" disabled={isSubmitting}>
-              {isSubmitting ? "Signing in..." : "Sign in"}
-            </Button>
-          </form>
-        </CardContent>
-        <CardFooter className="flex flex-col space-y-2">
-          <p className="text-muted-foreground text-center text-sm">
-            Don&apos;t have an account?{" "}
-            <Link
-              href="/auth/signup"
-              className="text-primary font-medium hover:underline"
-            >
-              Sign up
-            </Link>
-          </p>
-        </CardFooter>
-      </Card>
-    </div>
-  );
+  return <SignInForm />;
 }
